+++ src/internal/pkg/ydart/ydart.go
@@ -184,37 +184,37 @@ func (ydArt *YdArt) GenerateWithPrompt(prompt string, isDirectCall bool) (string
 	return response.Id, nil
 }
 
-func (ydArt *YdArt) GetImage(operationId string, filename string, filenameOriginalSize string) (bool, error) {
+func (ydArt *YdArt) GetImageSlice(operationId string) (bool, []byte, error) {
 	ydArt.logger.Debug("Get image request")
 	url := fmt.Sprintf("%s/operations/%s", CoreBaseURL, operationId)
 	var response getImageResponse
 	err := ydArt.innerRequest("GET", url, http.StatusOK, nil, &response)
 	if err != nil {
 		resultError := fmt.Errorf("error when get image: %v", err)
-		return false, resultError
+		return false, nil, resultError
 	}
 
 	if response.Done {
 		if response.Error != "" {
 			resultError := fmt.Errorf("error from YandexArt: %s", response.Error)
 			ydArt.logger.Error("YandexArt error", "errorCode", response.ErrorCode, "error", response.Error, "detail", response.ErrorDetails)
-			return true, resultError
+			return true, nil, resultError
 		}
 		if response.Response.Image != "" {
-			err := ydArt.ipr.ProcessImageFromBase64(filename, filenameOriginalSize, response.Response.Image, ydArt.imageParameters.Weight, ydArt.imageParameters.Height)
+			imageData, err := ydArt.ipr.ConvertBase64ToJpg(response.Response.Image)
 			if err != nil {
 				resultError := fmt.Errorf("error image processing: %v", err)
 				ydArt.logger.Error(resultError.Error())
-				return false, resultError
+				return false, nil, resultError
 			}
-			return true, nil
+			return true, imageData, nil
 		} else {
 			resultError := fmt.Errorf("field image is empty")
-			return false, resultError
+			return false, nil, resultError
 		}
 	}
 	// Задача не завершена
-	return false, nil
+	return false, nil, nil
 }
 
 func (ydArt *YdArt) IsReadyForRequest() bool {